Introduces a when before using goog.dependencies_, closes #143
Based on the reference at the bottom, goog.dependencies_ is available only if goog.COMPILED is not
true (according to David Nolen not available then for :simple and :advanced). In general the use of
goog.dependencies_ is not advisable as it is an implementation detail of the Google Closure Library.

The best option for including a lib is the :foreign-libs to read-eval-call.

Reference code:
https://google.github.io/closure-library/api/source/closure/goog/base.js.src.html#l795